pyproject.toml
[project]
name = "the-rals-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.5.0",
    "cohere>=5.15.0",
    "discord-py>=2.5.2",
    "groq>=0.24.0",
    "lancedb>=0.22.0",
    "mistralai>=1.7.0",
    "ollama>=0.4.8",
    "openai>=1.78.1",
    "pgvector>=0.4.1",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "tantivy>=0.24.0",
]
